net(kuri) : utilise uint64_t
"unsigned long" n'est pas garantie de mesure 8 octets sur toutes les
plateformes.

+++ logiciels/kuri/adn/genere_lexemes.cc
@@ -325,7 +325,7 @@ struct Lexeme {
     dls::vue_chaine_compacte chaine{};
 
 	union {
-		unsigned long long valeur_entiere;
+		uint64_t valeur_entiere;
 		double valeur_reelle;
 		long index_chaine;
 		IdentifiantCode *ident;
